break this step down for me:
In your /questions collection, each question document will need a new field, let's call it examIds. This field should be an array of strings. Each string in the array will be an ID of a practice test that this question belongs to.
Example: For dm_q1, if it's part of "UCAT DM Mock 1" (ID: ucat_dm_mock_1) and also a "Syllogisms Drill" (ID: dm_syllogisms_drill), the field would be: "examIds": ["ucat_dm_mock_1", "dm_syllogisms_drill"]